migrate from mysql to mongdb

+++ trunk/vineyard/internal/app/controllers/images_controller.rb
@@ -1,31 +1,56 @@
 class ImagesController < ApplicationController
-	# GET /resources/:id/images
+	# GET /products/:id/images
 	def index
+		count = Image.where(product_id: params[:product_id]).count
+		logger.info count
 		respond_to do |format|
-			format.json { render :json => 0}
+			if count == 0 
+				format.html {redirect_to action: 'new'}
+			else
+				@images = Image.where(product_id: params[:product_id])
+				logger.info @images
+				format.html # index.html.erb
+			end
+		end
+	end
+	
+	# GET /products/:id/images/new
+	def new
+		@product = Product.find(params[:product_id])
+		logger.info @product
+		respond_to do |format|
+			format.html # new.html.erb
 		end
 	end
 
-	# POST /resources/:id/images
+	# POST /products/:id/images
 	def create
-		image = Image.new({"photo" => params[:photo]})
-		image.resource_id = params[:resource_id].to_i
-		image.save
-		logger.info image
+		@image = Image.new(:product_id => params[:product_id])
+		@image.upload = params[:image][:upload]
+		@image.save!
 
 		respond_to do | format |
-			format.json { render :json => image }
+			format.json { render json: { :url => @image.upload.thumb.url, :id => @image.id} }
 		end
 	end
 
-	# DELETE /resources/:id/images/id
+	# GET /products/:id/images/:id
+	def show
+		
+		respond_to do |format|
+			format.html # show.html.erb
+		end
+	end
+
+	# DELETE /products/:id/images/id
 	def destroy
 		image = Image.find(params[:id])
-		image.photo.destroy
-		image.destroy
+		image.remove_upload! unless image == nil
+		image.destroy unless image == nil
 
 		respond_to do |format|
-			format.json{ render :json => 0 }
+			format.json { render :json => 0 }
+			format.html { redirect_to product_images_path}
 		end
 	end
 end
